Make PageReference a top level class
